Refactoring step 4: collect pg_dump functionality in pg_dump.go
Add pg_dump.go which contains:
- code from statement.go
- much of process.go (except for ProcessRow which moves to data.go)

Also migrate process_test.go to pgdump_test.go.
